+++ common/src/main/java/io/github/reserveword/imblocker/mixin/compat/MeteorTextFieldMixin.java
@@ -62,8 +62,8 @@ public abstract class MeteorTextFieldMixin extends MeteorWidgetMixin {
 	
 	@Override
 	public Point getCaretPos() {
-		int caretX = (int) (getTextWidth(cursor) - textStart + pad());
-		return new Point(getGuiScale(), caretX, (int) (height / 3.5));
+		double caretX = getTextWidth(cursor) - textStart + pad();
+		return new Point(getGuiScale(), caretX, height / 3.5);
 	}
 	
 	@Override
